Resolve unnecessary_cast clippy lint
    warning: casting to the same type is unnecessary (`u8` -> `u8`)
       --> src/lib.rs:404:25
        |
    404 |         data[raw_exp] = do_compute_exp_shift(bin_exp, dec_exp) as u8;
        |                         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `do_compute_exp_shift(bin_exp, dec_exp)`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
        = note: `-W clippy::unnecessary-cast` implied by `-W clippy::all`
        = help: to override `-W clippy::all` add `#[allow(clippy::unnecessary_cast)]`

+++ src/lib.rs
@@ -401,7 +401,7 @@ static EXP_SHIFTS: ExpShiftTable = {
             bin_exp += 1;
         }
         let dec_exp = compute_dec_exp(bin_exp, true);
-        data[raw_exp] = do_compute_exp_shift(bin_exp, dec_exp) as u8;
+        data[raw_exp] = do_compute_exp_shift(bin_exp, dec_exp);
         raw_exp += 1;
     }
